bot/requirements.txt
pytz
ujson
loguru
mariadb
pydantic
envparse
SQLAlchemy
python-telegram-bot[json]
python-telegram-bot-pagination
